style(table): moved some functions and added odd length compatibilty
