+++ src/documents/temp_interns.json
@@ -51,6 +51,13 @@
             "department": "Film Production",
             "location": "Watsonville"
 
+        },
+
+        {
+            "name": "Tinker Bell",
+            "department": "Film Production",
+            "location": "Gilroy"
+
         }
